[info panel] Some more refactoring and splitting of components, css
This is a step towards #288
