+++ editor/code/views/info/ErrorBrowser.tsx
@@ -1,0 +1,14 @@
+import { CoqPp } from "./CoqPp";
+import { Details } from "./Details";
+
+export type ErrorBrowserParams = { error?: string };
+
+export function ErrorBrowser({ error }: ErrorBrowserParams) {
+  if (!error) return null;
+
+  return (
+    <Details summary={"Error Browser"}>
+      <CoqPp content={error} inline={false} />
+    </Details>
+  );
+}
