+++ editor/code/views/info/media/index.css
